// core/enums/queryCategories.ts
enum EmailQueryCategories {
    REFUND = 'refund',
    DELIVERY_STATUS = 'delivery_status',
    SUBSCRIPTION_INQUIRY = 'subscription_inquiry',
    ORDER_CANCELLATION = 'order_cancellation',
    MULTIPLE_CONTEXTS = 'multiple_contexts',
    ESCALATION = 'escalation'
}


export default EmailQueryCategories;