// core/enums/statusTypes.ts
enum StatusTypes {
    PENDING = 'pending',
    ACCEPTED = 'accepted',
    DECLINED = 'declined',
    SCHEDULED = 'scheduled',
    STARTED = 'started',
    COMPLETED = 'completed',
    CANCELED = 'canceled',
    NOT_SCHEDULED = 'not_scheduled',
    APPROVED = 'approved',
    REJECTED = 'rejected',
    UNVERIFIED = 'unverified',
    VERIFIED = 'verified',
    GENERATED = 'generated',
    EXPIRED = 'expired',
    DONE = 'done',
    ACTIVE = 'active',
    INACTIVE = 'inactive',
    CREATED = 'created',
    DRAFT = 'draft',
    PUBLISHED = 'published',
    BOOKED = 'booked',
    RESCHEDULE_PENDING = 'reschedule_pending',
    RESCHEDULE_ACCEPTED = 'reschedule_accepted',
    FAILED = 'failed',
    SUCCESSFUL = 'successful',
    RESOLVED = 'resolved',
    RESTRICTED = 'restricted',
    BLOCKED = 'blocked',
    REVIEWED = 'reviewed',
    DELETED = 'deleted',
    REPORTED = 'reported',
    REMOVED = 'removed',
    IN_PROGRESS = 'in_progress',
    CLOSED = 'closed',
    PROCESSED = 'processed',
    DELIVERED = 'Delivered',
    UNPUBLISHED = 'unpublished',
}


export default StatusTypes;